+++ internal/sdkprovider/userconfig/serviceintegrationendpoint/external_google_cloud_bigquery.go
@@ -19,7 +19,7 @@ func externalGoogleCloudBigqueryUserConfig() *schema.Schema {
 				Type:        schema.TypeString,
 			},
 			"service_account_credentials": {
-				Description: "This is a JSON object with the fields documented in https://cloud.google.com/iam/docs/creating-managing-service-account-keys .",
+				Description: "This is a JSON object with the fields documented in https://cloud.google.com/iam/docs/creating-managing-service-account-keys.",
 				Required:    true,
 				Type:        schema.TypeString,
 			},